Fix Bugs

+++ libraries/src/tf2_ros/include/tf2_ros/static_transform_broadcaster_node.hpp
@@ -60,7 +60,7 @@ namespace tf2_ros
   {
   public:
     STATIC_TRANSFORM_BROADCASTER_PUBLIC
-    explicit StaticTransformBroadcasterNode(int16_t domain_id, TransformData transformData);
+    explicit StaticTransformBroadcasterNode(int domain_id, TransformData transformData);
 
     STATIC_TRANSFORM_BROADCASTER_PUBLIC
     ~StaticTransformBroadcasterNode() override = default;
